# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Detector API",
    "version": "0.1.0"
  },
  "paths": {
    "/health/live": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Live",
        "operationId": "live_health_live_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LivenessRead"
                }
              }
            }
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Ready",
        "operationId": "ready_health_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessRead"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessRead"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/system/status": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Get System Status",
        "operationId": "get_system_status_api_v1_system_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatusRead"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cameras": {
      "get": {
        "tags": [
          "cameras"
        ],
        "summary": "List Cameras",
        "operationId": "list_cameras_api_v1_cameras_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CameraRead"
                  },
                  "type": "array",
                  "title": "Response List Cameras Api V1 Cameras Get"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "cameras"
        ],
        "summary": "Create Camera",
        "operationId": "create_camera_api_v1_cameras_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CameraCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CameraRead"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cameras/{camera_id}": {
      "get": {
        "tags": [
          "cameras"
        ],
        "summary": "Get Camera",
        "operationId": "get_camera_api_v1_cameras__camera_id__get",
        "parameters": [
          {
            "name": "camera_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Camera Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CameraRead"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          }
        }
      },
      "patch": {
        "tags": [
          "cameras"
        ],
        "summary": "Update Camera",
        "operationId": "update_camera_api_v1_cameras__camera_id__patch",
        "parameters": [
          {
            "name": "camera_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Camera Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CameraUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CameraRead"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          }
        }
      },
      "delete": {
        "tags": [
          "cameras"
        ],
        "summary": "Delete Camera",
        "operationId": "delete_camera_api_v1_cameras__camera_id__delete",
        "parameters": [
          {
            "name": "camera_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Camera Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/cameras/{camera_id}/start": {
      "post": {
        "tags": [
          "cameras"
        ],
        "summary": "Start Camera",
        "operationId": "start_camera_api_v1_cameras__camera_id__start_post",
        "parameters": [
          {
            "name": "camera_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Camera Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CameraRead"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/cameras/{camera_id}/stop": {
      "post": {
        "tags": [
          "cameras"
        ],
        "summary": "Stop Camera",
        "operationId": "stop_camera_api_v1_cameras__camera_id__stop_post",
        "parameters": [
          {
            "name": "camera_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Camera Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CameraRead"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/cameras/{camera_id}/restart": {
      "post": {
        "tags": [
          "cameras"
        ],
        "summary": "Restart Camera",
        "operationId": "restart_camera_api_v1_cameras__camera_id__restart_post",
        "parameters": [
          {
            "name": "camera_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Camera Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CameraRead"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/uploads": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Upload Video",
        "operationId": "upload_video_api_v1_uploads_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_video_api_v1_uploads_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadRead"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_video_api_v1_uploads_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_video_api_v1_uploads_post"
      },
      "CameraCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceType"
          },
          "source_url": {
            "type": "string",
            "maxLength": 2048,
            "title": "Source Url"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "source_type",
          "source_url"
        ],
        "title": "CameraCreate"
      },
      "CameraRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceType"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url",
            "description": "Credentials are masked"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "Desired state: the camera should be running"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "runtime": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CameraRuntimeRead"
              },
              {
                "type": "null"
              }
            ],
            "description": "Live status from the worker; null when no worker reports this camera"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "source_type",
          "source_url",
          "enabled",
          "created_at",
          "updated_at",
          "runtime"
        ],
        "title": "CameraRead"
      },
      "CameraRuntimeRead": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/CameraStatus"
          },
          "capture_fps": {
            "type": "number",
            "title": "Capture Fps"
          },
          "frame_size": {
            "anyOf": [
              {
                "prefixItems": [
                  {
                    "type": "integer"
                  },
                  {
                    "type": "integer"
                  }
                ],
                "type": "array",
                "maxItems": 2,
                "minItems": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Frame Size"
          },
          "last_frame_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Frame At"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "reconnect_attempts": {
            "type": "integer",
            "title": "Reconnect Attempts"
          },
          "last_reconnect_attempt_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Reconnect Attempt At"
          },
          "worker_id": {
            "type": "string",
            "title": "Worker Id"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "status",
          "capture_fps",
          "frame_size",
          "last_frame_at",
          "last_error",
          "reconnect_attempts",
          "last_reconnect_attempt_at",
          "worker_id",
          "updated_at"
        ],
        "title": "CameraRuntimeRead"
      },
      "CameraStatus": {
        "type": "string",
        "enum": [
          "CREATED",
          "STARTING",
          "RUNNING",
          "DEGRADED",
          "RECONNECTING",
          "STOPPING",
          "STOPPED",
          "ERROR"
        ],
        "title": "CameraStatus"
      },
      "CameraUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "source_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceType"
              },
              {
                "type": "null"
              }
            ]
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CameraUpdate"
      },
      "DependencyStatusRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "name",
          "ok",
          "error"
        ],
        "title": "DependencyStatusRead"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "LivenessRead": {
        "properties": {
          "status": {
            "type": "string",
            "const": "alive",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "LivenessRead"
      },
      "ReadinessRead": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "not_ready"
            ],
            "title": "Status"
          },
          "dependencies": {
            "items": {
              "$ref": "#/components/schemas/DependencyStatusRead"
            },
            "type": "array",
            "title": "Dependencies"
          }
        },
        "type": "object",
        "required": [
          "status",
          "dependencies"
        ],
        "title": "ReadinessRead"
      },
      "SourceType": {
        "type": "string",
        "enum": [
          "file",
          "rtsp",
          "mock"
        ],
        "title": "SourceType"
      },
      "SystemStatusRead": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "degraded"
            ],
            "title": "Status"
          },
          "checked_at": {
            "type": "string",
            "format": "date-time",
            "title": "Checked At"
          },
          "dependencies": {
            "items": {
              "$ref": "#/components/schemas/DependencyStatusRead"
            },
            "type": "array",
            "title": "Dependencies"
          },
          "workers": {
            "items": {
              "$ref": "#/components/schemas/WorkerStatusRead"
            },
            "type": "array",
            "title": "Workers"
          }
        },
        "type": "object",
        "required": [
          "status",
          "checked_at",
          "dependencies",
          "workers"
        ],
        "title": "SystemStatusRead"
      },
      "UploadRead": {
        "properties": {
          "file_ref": {
            "type": "string",
            "title": "File Ref"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "original_name": {
            "type": "string",
            "title": "Original Name"
          }
        },
        "type": "object",
        "required": [
          "file_ref",
          "source_url",
          "size_bytes",
          "original_name"
        ],
        "title": "UploadRead"
      },
      "WorkerStatusRead": {
        "properties": {
          "worker_id": {
            "type": "string",
            "title": "Worker Id"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "last_seen_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen At"
          },
          "uptime_seconds": {
            "type": "number",
            "title": "Uptime Seconds"
          }
        },
        "type": "object",
        "required": [
          "worker_id",
          "version",
          "started_at",
          "last_seen_at",
          "uptime_seconds"
        ],
        "title": "WorkerStatusRead"
      }
    }
  }
}